+++ comics/comics/applegeekslite.py
@@ -6,16 +6,10 @@ class Meta(MetaBase):
     language = 'en'
     url = 'http://www.applegeeks.com/'
     start_date = '2006-04-18'
+    end_date = '2010-08-30'
+    active = False
     rights = 'Mohammad Haque & Ananth Panagariya'
 
 class Crawler(CrawlerBase):
-    history_capable_days = 30
-    schedule = None
-    time_zone = -5
-
     def crawl(self, pub_date):
-        feed = self.parse_feed('http://www.applegeeks.com/rss/?cat=lite')
-        for entry in feed.for_date(pub_date):
-            url = entry.summary.src('img')
-            title = entry.title.replace('AG Lite - ', '')
-            return CrawlerImage(url, title)
+        pass # Comic no longer published